this is a small renaming + adding the data folder, bringin in all notebooks related resilience atlas
