--- cloud_functions/.editorconfig
@@ -1,36 +1,0 @@
-# editorconfig.org
-
-root = true
-
-[*]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-
-[*.md]
-indent_style = tab
-indent_size = 4
-trim_trailing_whitespace = false
-
-[*.yml]
-indent_style = space
-indent_size = 2
-
-[*.yaml]
-indent_style = space
-indent_size = 2
-
-[*.json]
-indent_style = space
-indent_size = 2
-
-[*.js]
-indent_style = tab
-indent_size = 4
-
-[*.py]
-indent_style = tab
-indent_size = 4
