+++ cloud_functions/README.md
@@ -2,13 +2,13 @@
 
 This repository contains the AWS Lambda functions and google cloud functions used by the Resilience Atlas.
 
-cloud_functions  
- ┣ download_image  
- ┣ histogram  
- ┣ raster_interaction  
- ┣ sparc_plots  
- ┣ titiler_cogs  
- ┣ .editorconfig  
- ┗ README.md  
+cloud_functions
+ ┣ analysis_categorical
+ ┣ analysis_histogram
+ ┣ download_image
+ ┣ raster_interaction
+ ┣ sparc_plots
+ ┣ titiler_cogs
+ ┗ README.md
 
 Instructions on how to develop and deploy the functions are in the README.md files in each folder.
