--- cloud_functions/histogram/.gitignore
@@ -1,7 +1,0 @@
-*/node_modules
-.DS_STORE
-npm-debug.log
-prod.env
-*.env
-package-lock.json
-*privatekey.json
