--- cloud_functions/histogram/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "rasterInteraction",
-  "version": "0.0.1",
-  "private": true,
-  "license": "MIT",
-  "author": "Alicia Arenzana <alicia.arenzana@vizzuality.com>",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/GoogleCloudPlatform/nodejs-docs-samples.git"
-  },
-  "engines": {
-    "node": ">=8"
-  },
-  "dependencies": {
-    "@google/earthengine": "^0.1.232"
-  },
-  "cloud-repo-tools": {
-    "requiresKeyFile": true,
-    "requiresProjectId": true
-  },
-  "scripts": {
-    "start": "functions-framework --target=rasterInteraction"
-  },
-  "bugs": {
-    "url": "https://github.com/GoogleCloudPlatform/nodejs-docs-samples/issues"
-  },
-  "homepage": "https://github.com/GoogleCloudPlatform/nodejs-docs-samples#readme",
-  "main": "index.js",
-  "devDependencies": {
-    "@google-cloud/functions-framework": "^1.2.1"
-  },
-  "description": "this is a test"
-}
