+++ data/.pre-commit-config.yaml
@@ -1,0 +1,26 @@
+repos:
+    -   repo: https://github.com/pre-commit/pre-commit-hooks
+        rev: v2.3.0
+        hooks:
+        -   id: check-yaml
+        -   id: end-of-file-fixer
+        -   id: trailing-whitespace
+    # - repo: meta
+    #     hooks:
+    #       - id: check-hooks-apply
+    #       - id: check-useless-excludes
+
+    -   repo: https://github.com/psf/black
+        rev: 22.10.0
+        hooks:
+        - id: black-jupyter
+        language_version: python3.9
+            # It is recommended to specify the latest version of Python
+            # supported by your project here, or alternatively use
+            # pre-commit's default_language_version, see
+            # https://pre-commit.com/#top_level-default_language_version
+
+    -   repo: https://gitlab.com/PyCQA/flake8
+        rev: 3.8.3
+        hooks:
+        - id: flake8
