+++ data/notebooks/jupyter_server_config.py
@@ -1,0 +1,67 @@
+from jupyter_core.paths import jupyter_data_dir
+import subprocess
+import os
+import stat
+
+# https://jupyter-notebook.readthedocs.io/en/stable/config.html
+c = get_config()  # noqa: F821
+c.ServerApp.ip = '0.0.0.0'
+c.ServerApp.port = 8888
+c.ServerApp.open_browser = False
+c.ServerApp.disable_check_xsrf = True
+# to avoid having to auth with jupyter
+c.ServerApp.token = ''
+c.ServerApp.password = ''
+
+# https://github.com/jupyter/notebook/issues/3130
+c.FileContentsManager.delete_to_trash = False
+
+# To output both image/svg+xml and application/pdf plot formats
+# in the notebook file
+c.InlineBackend.figure_formats = {"png", "jpeg", "svg", "pdf"}
+
+# Generate a self-signed certificate
+OPENSSL_CONFIG = """\
+[req]
+distinguished_name = req_distinguished_name
+[req_distinguished_name]
+"""
+if "GEN_CERT" in os.environ:
+    dir_name = jupyter_data_dir()
+    pem_file = os.path.join(dir_name, "notebook.pem")
+    os.makedirs(dir_name, exist_ok=True)
+
+    # Generate an openssl.cnf file to set the distinguished name
+    cnf_file = os.path.join(os.getenv("CONDA_DIR", "/usr/lib"), "ssl",
+                            "openssl.cnf")
+    if not os.path.isfile(cnf_file):
+        with open(cnf_file, "w") as fh:
+            fh.write(OPENSSL_CONFIG)
+
+    # Generate a certificate if one doesn't exist on disk
+    subprocess.check_call([
+        "openssl",
+        "req",
+        "-new",
+        "-newkey=rsa:2048",
+        "-days=365",
+        "-nodes",
+        "-x509",
+        "-subj=/C=XX/ST=XX/L=XX/O=generated/CN=generated",
+        f"-keyout={pem_file}",
+        f"-out={pem_file}",
+    ])
+    # Restrict access to the file
+    os.chmod(pem_file, stat.S_IRUSR | stat.S_IWUSR)
+    c.ServerApp.certfile = pem_file
+
+# Change default umask for all subprocesses of the notebook server if set in
+# the environment
+if "NB_UMASK" in os.environ:
+    os.umask(int(os.environ["NB_UMASK"], 8))
+
+# make template directory
+# c.JupyterLabTemplates.template_dirs = [
+# '/home/jovyan/work/notebooks/templates']
+# c.JupyterLabTemplates.include_default = True
+# c.JupyterLabTemplates.include_core_paths = True
